requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
langgraph>=0.1.0
python-dotenv>=1.0.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
pytest>=8.0.0